Вынес общую логику bold и italic handlerов
